Add --apply-differences-from= option to 'ostree checkout'
If we already have an old checkout at a known revision, for small
changes, it can be much faster to modify that than to checkout an
entire new tree of hardlinks. Add an --apply-differences-from=
argument to 'ostree checkout' that says to assume that the
destination is already at the given commit and do an incremental
update.